Refactor in app.rb
